+++ Changelog.md
@@ -4,3 +4,4 @@
 ## v0.2.2
 
 * Started a Changelog, let's see how long I keep it going.
+* Added read only and write only variants.